P2/web/requirements.txt
flask
flask_wtf
pyjwt
redis
requests
python-dotenv